requirements.txt
pydantic==1.10.8
PyYAML==6.0
superagi-tools==1.0.6
requests==2.31.0
retry==0.9.2
setuptools==63.2.0
superagi-tools==1.0.6
trafilatura==1.6.2